+++ CHANGELOG.rst
@@ -27,6 +27,8 @@ Changes:
 - Added ``AppConfig.from_environ()`` to instantiate the config class without importing the ``environ`` module.
 - Added ``AppConfig.generate_help()`` to generate the help without a need to import the ``environ`` module.
 - Added ``environ.config(from_environ="fr_env", generate_help="gen_hlp")`` class methods can be renamed or prevented
+- If ``environ.var`` is passed an ``attr.Factory``, the callable is used to generate the default value.
+
 
 
 ----